Lots of Arrows
Did lots of changes in regards to the MRV overlay. There was a lot of wondering as to why different methods of getting a "resultant vector" had very different outcomes when it came to differing amounts of bins... Something to think about... You know? I feel like with more intuition and digging and testing, one could write a small paper out of this... hmmm