Bandaid
So I added some functionality that scales the LR and HR plots to look definitely different. The axis is now fixed, so you can choose it. Anyways, the weird ass time shift fucked me up, so I manually shifted it by 5 hours to fix it... The actual actigraphy that leads to the fucked up values is not fixed though... I gotta do some more digging is all...